Organize unit tests by type

moved a few more tests out
